@@ -15,6 +15,7 @@ using Microsoft.Diagnostics.Tracing.Parsers.InteropEventProvider;
 using Microsoft.Diagnostics.Tracing.Parsers.JSDumpHeap;
 using Microsoft.Diagnostics.Tracing.Parsers.Kernel;
 using Microsoft.Diagnostics.Tracing.Stacks;
+using Microsoft.Diagnostics.Tracing.StackSources;
 using Microsoft.Diagnostics.Utilities;
 using Microsoft.DiagnosticsHub.Packaging.InteropEx;
 using PerfView.GuiUtilities;
@@ -157,7 +158,7 @@ namespace PerfView
                                     continue;
                                 }
                                 // We know that .NGENPDB directories are uninteresting, filter them out.  
-                                if (dir.EndsWith(".NENPDB", StringComparison.OrdinalIgnoreCase))
+                                if (dir.EndsWith(".NGENPDB", StringComparison.OrdinalIgnoreCase))
                                 {
                                     continue;
                                 }
@@ -2668,7 +2669,7 @@ table {
                 }
             };
 
-            dispatcher.Clr.ContentionStart += delegate (ContentionTraceData data)
+            dispatcher.Clr.ContentionStart += delegate (ContentionStartTraceData data)
             {
                 int idx = GetBucket(data.TimeStampRelativeMSec, startIntervalMSec, bucketIntervalMSec, byTimeStats.Length);
                 if (idx >= 0)
@@ -4177,15 +4178,18 @@ table {
         protected internal override StackSource OpenStackSourceImpl(string streamName, TextWriter log, double startRelativeMSec = 0, double endRelativeMSec = double.PositiveInfinity, Predicate<TraceEvent> predicate = null)
         {
             var eventLog = GetTraceLog(log);
-            if (streamName == "CPU")
+            bool showOptimizationTiers =
+                App.CommandLineArgs.ShowOptimizationTiers || streamName.Contains("(with Optimization Tiers)");
+            if (streamName.StartsWith("CPU"))
             {
-                return eventLog.CPUStacks(null, App.CommandLineArgs.ShowUnknownAddresses, predicate);
+                return eventLog.CPUStacks(null, App.CommandLineArgs, showOptimizationTiers, predicate);
             }
 
             // var stackSource = new InternTraceEventStackSource(eventLog);
             var stackSource = new MutableTraceEventStackSource(eventLog);
 
             stackSource.ShowUnknownAddresses = App.CommandLineArgs.ShowUnknownAddresses;
+            stackSource.ShowOptimizationTiers = showOptimizationTiers;
 
             TraceEvents events = eventLog.Events;
             if (!streamName.Contains("TaskTree") && !streamName.Contains("Tasks)"))
@@ -6575,7 +6579,7 @@ table {
 
         protected internal override void ConfigureStackWindow(string stackSourceName, StackWindow stackWindow)
         {
-            ConfigureAsEtwStackWindow(stackWindow, stackSourceName == "CPU");
+            ConfigureAsEtwStackWindow(stackWindow, stackSourceName.StartsWith("CPU"));
 
             if (stackSourceName == "Processes / Files / Registry")
             {
@@ -6608,7 +6612,7 @@ table {
                 stackWindow.ExcludeRegExTextBox.Text = excludePat;
             }
 
-            if (stackSourceName == "CPU" || stackSourceName.Contains("Thread Time"))
+            if (stackSourceName.StartsWith("CPU") || stackSourceName.Contains("Thread Time"))
             {
                 if (m_traceLog != null)
                 {
@@ -6991,6 +6995,13 @@ table {
             if (hasCPUStacks)
             {
                 m_Children.Add(new PerfViewStackSource(this, "CPU"));
+
+                if (!App.CommandLineArgs.ShowOptimizationTiers &&
+                    tracelog.Events.Any(
+                        e => e is MethodLoadUnloadTraceDataBase td && td.OptimizationTier != OptimizationTier.Unknown))
+                {
+                    advanced.Children.Add(new PerfViewStackSource(this, "CPU (with Optimization Tiers)"));
+                }
             }
 
             if (hasCSwitchStacks)
@@ -7225,6 +7236,10 @@ table {
                 options.KeepAllEvents = true;
             }
 
+            var traceEventDispatcherOptions = new TraceEventDispatcherOptions();
+            traceEventDispatcherOptions.StartTime = App.CommandLineArgs.StartTime;
+            traceEventDispatcherOptions.EndTime = App.CommandLineArgs.EndTime;
+
             options.MaxEventCount = App.CommandLineArgs.MaxEventCount;
             options.ContinueOnError = App.CommandLineArgs.ContinueOnError;
             options.SkipMSec = App.CommandLineArgs.SkipMSec;
@@ -7249,7 +7264,7 @@ table {
                 if (!File.Exists(etlxFile))
                 {
                     log.WriteLine("Creating ETLX file {0} from {1}", etlxFile, dataFileName);
-                    TraceLog.CreateFromEventTraceLogFile(dataFileName, etlxFile, options);
+                    TraceLog.CreateFromEventTraceLogFile(dataFileName, etlxFile, options, traceEventDispatcherOptions);
 
                     var dataFileSize = "Unknown";
                     if (File.Exists(dataFileName))
@@ -8250,6 +8265,7 @@ table {
         private string[] PerfScriptStreams = new string[]
         {
             "CPU",
+            "CPU (with Optimization Tiers)",
             "Thread Time (experimental)"
         };
 
@@ -8303,7 +8319,9 @@ table {
                         new ParallelLinuxPerfScriptStackSource(FilePath, doThreadTime), xmlPath);
                 }
 
-                return new XmlStackSource(xmlPath);
+                bool showOptimizationTiers =
+                    App.CommandLineArgs.ShowOptimizationTiers || streamName.Contains("(with Optimization Tiers)");
+                return new XmlStackSource(xmlPath, null, showOptimizationTiers);
             }
 
             return null;
@@ -8336,6 +8354,15 @@ table {
             var memory = new PerfViewTreeGroup("Memory Group");
 
             m_Children.Add(new PerfViewStackSource(this, "CPU"));
+
+            if (!App.CommandLineArgs.ShowOptimizationTiers &&
+                m_traceLog != null &&
+                m_traceLog.Events.Any(
+                    e => e is MethodLoadUnloadTraceDataBase td && td.OptimizationTier != OptimizationTier.Unknown))
+            {
+                advanced.AddChild(new PerfViewStackSource(this, "CPU (with Optimization Tiers)"));
+            }
+
             if (AppLog.InternalUser)
             {
                 advanced.AddChild(new PerfViewStackSource(this, "Thread Time (experimental)"));
@@ -8617,6 +8644,7 @@ table {
                         stackSource.OnlyManagedCodeStacks = true;
 
                         stackSource.ShowUnknownAddresses = App.CommandLineArgs.ShowUnknownAddresses;
+                        stackSource.ShowOptimizationTiers = App.CommandLineArgs.ShowOptimizationTiers;
 
                         TraceEvents events = eventLog.Events;
 
