@@ -898,7 +898,7 @@ namespace PerfView
                 saveDialog.InitialDirectory = Path.GetDirectoryName(DataSource.FilePath);
                 saveDialog.Title = "File to save view";
                 saveDialog.DefaultExt = ".perfView.xml.zip";
-                saveDialog.Filter = "PerfView view file|*.perfView.xml.zip|Comma Separated Value|*.csv|Speed Scope Format|*.speedscope.json|All Files|*.*";
+                saveDialog.Filter = "PerfView view file|*.perfView.xml.zip|Comma Separated Value|*.csv|Speed Scope|*.speedscope.json|Chromium Trace Event|*.chromium.json|All Files|*.*";
                 saveDialog.AddExtension = true;
                 saveDialog.OverwritePrompt = true;
 
@@ -947,6 +947,10 @@ namespace PerfView
             else if(m_fileName.EndsWith(".speedscope.json", StringComparison.OrdinalIgnoreCase))
             {
                 SpeedScopeStackSourceWriter.WriteStackViewAsJson(CallTree.StackSource, m_fileName);
+            }
+            else if (m_fileName.EndsWith(".chromium.json", StringComparison.OrdinalIgnoreCase))
+            {
+                ChromiumStackSourceWriter.WriteStackViewAsJson(CallTree.StackSource, m_fileName, false);
             }
             else
             {
