@@ -1,5 +1,6 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+﻿// <copyright file="LineInformation.cs" company="Microsoft">
+// Copyright (c) Microsoft. All rights reserved.
+// </copyright>
 
 namespace PerfViewJS
 {
@@ -8,7 +9,5 @@ namespace PerfViewJS
         public int LineNumber { get; set; }
 
         public double Metric { get; set; }
-
-        public string Line { get; set; }
     }
 }
