@@ -1,5 +1,6 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+﻿// <copyright file="SourceInformation.cs" company="Microsoft">
+// Copyright (c) Microsoft. All rights reserved.
+// </copyright>
 
 namespace PerfViewJS
 {
@@ -10,16 +11,16 @@ namespace PerfViewJS
     public sealed class SourceInformation
     {
         [DataMember]
-        public string Type { get; set; }
-
-        [DataMember]
         public string Url { get; set; }
 
         [DataMember]
-        public IEnumerable<LineInformation> Lines { get; set; }
+        public string Log { get; set; }
 
         [DataMember]
         public IEnumerable<LineInformation> Summary { get; set; }
+
+        [DataMember]
+        public string Data { get; set; }
 
         [DataMember]
         public string BuildTimeFilePath { get; set; }
