@@ -514,8 +514,12 @@ namespace Diagnostics.Tracing.StackSources
         {
             if (recursionGuard.RequiresNewThread)
             {
+                // Avoid capturing method parameters for use in the lambda to reduce fast-path allocation costs
+                var capturedThis = this;
+                var capturedStackIndex = stackIndex;
+                var capturedRecursionGuard = recursionGuard;
                 Task<StackInfo> operation = Task.Factory.StartNew(
-                    () => GetStackInfo(stackIndex, recursionGuard.ResetOnNewThread),
+                    () => capturedThis.GetStackInfo(capturedStackIndex, capturedRecursionGuard.ResetOnNewThread),
                     TaskCreationOptions.LongRunning);
 
                 return operation.GetAwaiter().GetResult();
