@@ -56,14 +56,14 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
         /// <para>If etlxFilePath is null the output name is derived from etlFilePath by changing its file extension to .ETLX.</para>
         /// <returns>The name of the ETLX file that was generated.</returns>
         /// </summary>
-        public static string CreateFromEventTraceLogFile(string filePath, string etlxFilePath = null, TraceLogOptions options = null)
+        public static string CreateFromEventTraceLogFile(string filePath, string etlxFilePath = null, TraceLogOptions options = null, TraceEventDispatcherOptions traceEventDispatcherOptions = null)
         {
             if (etlxFilePath == null)
             {
                 etlxFilePath = Path.ChangeExtension(filePath, ".etlx");
             }
 
-            using (TraceEventDispatcher source = TraceEventDispatcher.GetDispatcherFromFileName(filePath))
+            using (TraceEventDispatcher source = TraceEventDispatcher.GetDispatcherFromFileName(filePath, traceEventDispatcherOptions))
             {
                 if (source.EventsLost != 0 && options != null && options.OnLostEvents != null)
                 {
@@ -1338,9 +1338,9 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
 
             var ClrRundownParser = new ClrRundownTraceEventParser(rawEvents);
             Action<ModuleLoadUnloadTraceData> onLoaderRundown = delegate (ModuleLoadUnloadTraceData data)
-           {
-               processes.GetOrCreateProcess(data.ProcessID, data.TimeStampQPC).LoadedModules.ManagedModuleLoadOrUnload(data, false, true);
-           };
+            {
+                processes.GetOrCreateProcess(data.ProcessID, data.TimeStampQPC).LoadedModules.ManagedModuleLoadOrUnload(data, false, true);
+            };
 
             ClrRundownParser.LoaderModuleDCStop += onLoaderRundown;
             ClrRundownParser.LoaderModuleDCStart += onLoaderRundown;
@@ -2135,10 +2135,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                         // simulate a module unload, and resolve all code addresses in the module's range.   
                         CodeAddresses.ForAllUnresolvedCodeAddressesInRange(process, module.ImageBase, module.ModuleFile.ImageSize, false, delegate (ref TraceCodeAddresses.CodeAddressInfo info)
                         {
-                            if (info.moduleFileIndex == Microsoft.Diagnostics.Tracing.Etlx.ModuleFileIndex.Invalid)
-                            {
-                                info.moduleFileIndex = module.ModuleFile.ModuleFileIndex;
-                            }
+                            info.SetModuleFileIndex(module.ModuleFile);
                         });
                     }
                     if (module.unloadTimeQPC > sessionEndTimeQPC)
@@ -6606,14 +6603,10 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                 }
 
                 // Look for all code addresses those that don't have modules that are in my range are assumed to be mine.  
-                var moduleFileIndex = module.ModuleFile.ModuleFileIndex;
                 Process.Log.CodeAddresses.ForAllUnresolvedCodeAddressesInRange(process, data.ImageBase, data.ImageSize, false,
                     delegate (ref Microsoft.Diagnostics.Tracing.Etlx.TraceCodeAddresses.CodeAddressInfo info)
                     {
-                        if (info.moduleFileIndex == Microsoft.Diagnostics.Tracing.Etlx.ModuleFileIndex.Invalid)
-                        {
-                            info.moduleFileIndex = moduleFileIndex;
-                        }
+                        info.SetModuleFileIndex(moduleFile);
                     });
             }
             CheckClassInvarients();
@@ -6695,6 +6688,8 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                     module.NativeModule.ModuleFile.pdbAge = data.NativePdbAge;
                     module.NativeModule.ModuleFile.pdbName = data.NativePdbBuildPath;
                 }
+
+                module.InitializeNativeModuleIsReadyToRun();
             }
 
             // TODO factor this with the unmanaged case.  
@@ -7151,6 +7146,14 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
             : base(process, moduleFile, moduleID)
         { }
 
+        internal void InitializeNativeModuleIsReadyToRun()
+        {
+            if (NativeModule != null && (flags & ModuleFlags.ReadyToRunModule) != ModuleFlags.None)
+            {
+                NativeModule.ModuleFile.isReadyToRun = true;
+            }
+        }
+
         // TODO use or remove
         internal TraceLoadedModule nativeModule;        // non-null for IL managed modules
         internal long assemblyID;
@@ -7170,6 +7173,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
             deserializer.Read(out assemblyID);
             deserializer.Read(out nativeModule);
             deserializer.Read(out flags); this.flags = (ModuleFlags)flags;
+            InitializeNativeModuleIsReadyToRun();
         }
         #endregion
     }
@@ -7572,9 +7576,15 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
     public sealed class TraceCodeAddresses : IFastSerializable, IEnumerable<TraceCodeAddress>
     {
         /// <summary>
+        /// Chunk size for <see cref="codeAddressObjects"/>
+        /// </summary>
+        private const int ChunkSize = 4096;
+
+        /// <summary>
         /// Returns the count of code address indexes (all code address indexes are strictly less than this).   
         /// </summary>
         public int Count { get { return codeAddresses.Count; } }
+
         /// <summary>
         /// Given a code address index, return the name associated with it (the method name).  It will
         /// have the form MODULE!METHODNAME.   If the module name is unknown a ? is used, and if the
@@ -7582,12 +7592,13 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
         /// </summary>
         public string Name(CodeAddressIndex codeAddressIndex)
         {
-            if (names == null)
+            if (this.names == null)
             {
-                names = new string[Count];
+                this.names = new string[Count];
             }
 
-            string name = names[(int)codeAddressIndex];
+            string name = this.names[(int)codeAddressIndex];
+
             if (name == null)
             {
                 string moduleName = "?";
@@ -7608,8 +7619,9 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                     methodName = "0x" + ((ulong)Address(codeAddressIndex)).ToString("x");
                 }
 
-                name = moduleName + "!" + methodName;
+                this.names[(int)codeAddressIndex] = name = moduleName + "!" + methodName;
             }
+
             return name;
         }
         /// <summary>
@@ -7654,6 +7666,13 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
 
             return ilMap.GetILOffsetForNativeAddress(Address(codeAddressIndex));
         }
+
+        public OptimizationTier OptimizationTier(CodeAddressIndex codeAddressIndex)
+        {
+            Debug.Assert((int)codeAddressIndex < codeAddresses.Count);
+            return codeAddresses[(int)codeAddressIndex].optimizationTier;
+        }
+
         /// <summary>
         /// Given a code address index, returns a TraceCodeAddress for it.
         /// </summary>
@@ -7661,22 +7680,38 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
         {
             get
             {
-                if (codeAddressObjects == null || (int)codeAddressIndex >= codeAddressObjects.Length)
-                {
-                    codeAddressObjects = new TraceCodeAddress[(int)codeAddressIndex + 16];
-                }
-
                 if (codeAddressIndex == CodeAddressIndex.Invalid)
                 {
                     return null;
                 }
 
-                TraceCodeAddress ret = codeAddressObjects[(int)codeAddressIndex];
+                int chunk = (int)codeAddressIndex / ChunkSize;
+                int offset = (int)codeAddressIndex % ChunkSize;
+
+                if (this.codeAddressObjects == null)
+                {
+                    this.codeAddressObjects = new TraceCodeAddress[chunk + 1][];
+                }
+                else if (chunk >= this.codeAddressObjects.Length)
+                {
+                    Array.Resize(ref this.codeAddressObjects, Math.Max(this.codeAddressObjects.Length * 2, chunk + 1));
+                }
+
+                TraceCodeAddress[] data = this.codeAddressObjects[chunk];
+
+                if (data == null)
+                {
+                    data = this.codeAddressObjects[chunk] = new TraceCodeAddress[ChunkSize];
+                }
+
+                TraceCodeAddress ret = data[offset];
+
                 if (ret == null)
                 {
                     ret = new TraceCodeAddress(this, codeAddressIndex);
-                    codeAddressObjects[(int)codeAddressIndex] = ret;
+                    data[offset] = ret;
                 }
+
                 return ret;
             }
         }
@@ -8031,6 +8066,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                     {
                         info.SetILMapIndex(this, ilMap);
                     }
+                    info.SetOptimizationTier(data.OptimizationTier);
                 }
             });
         }
@@ -8440,7 +8476,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
             // If we have a signature, use it
             if (moduleFile.PdbSignature != Guid.Empty)
             {
-                pdbFileName = symReader.FindSymbolFilePath(moduleFile.PdbName, moduleFile.PdbSignature, moduleFile.PdbAge, moduleFile.FilePath, moduleFile.ProductVersion);
+                pdbFileName = symReader.FindSymbolFilePath(moduleFile.PdbName, moduleFile.PdbSignature, moduleFile.PdbAge, moduleFile.FilePath, moduleFile.ProductVersion, true);
             }
             else
             {
@@ -8559,6 +8595,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                 serializer.Write(moduleFiles);
                 serializer.Write(methods);
 
+                serializer.WriteTagged(CodeAddressInfoSerializationVersion);
                 serializer.Write(codeAddresses.Count);
                 serializer.Log("<WriteCollection name=\"codeAddresses\" count=\"" + codeAddresses.Count + "\">\r\n");
                 for (int i = 0; i < codeAddresses.Count; i++)
@@ -8567,6 +8604,9 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                     serializer.Write((int)codeAddresses[i].moduleFileIndex);
                     serializer.Write((int)codeAddresses[i].methodOrProcessOrIlMapIndex);
                     serializer.Write(codeAddresses[i].InclusiveCount);
+
+                    /// <see cref="CodeAddressInfoSerializationVersion"/> >= 1
+                    serializer.Write((byte)codeAddresses[i].optimizationTier);
                 }
                 serializer.Write(totalCodeAddresses);
                 serializer.Log("</WriteCollection>\r\n");
@@ -8587,6 +8627,8 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                 deserializer.Read(out moduleFiles);
                 deserializer.Read(out methods);
 
+                int storedCodeAddressInfoSerializationVersion = 0;
+                deserializer.TryReadTagged(ref storedCodeAddressInfoSerializationVersion);
                 int count = deserializer.ReadInt();
                 deserializer.Log("<Marker name=\"codeAddresses\" count=\"" + count + "\"/>");
                 CodeAddressInfo codeAddressInfo = new CodeAddressInfo();
@@ -8597,6 +8639,12 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                     codeAddressInfo.moduleFileIndex = (ModuleFileIndex)deserializer.ReadInt();
                     codeAddressInfo.methodOrProcessOrIlMapIndex = deserializer.ReadInt();
                     deserializer.Read(out codeAddressInfo.InclusiveCount);
+
+                    if (storedCodeAddressInfoSerializationVersion >= 1)
+                    {
+                        codeAddressInfo.optimizationTier = (OptimizationTier)deserializer.ReadByte();
+                    }
+
                     codeAddresses.Add(codeAddressInfo);
                 }
                 deserializer.Read(out totalCodeAddresses);
@@ -8609,6 +8657,8 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
             });
             lazyCodeAddresses.FinishRead();        // TODO REMOVE 
         }
+
+        private const int CodeAddressInfoSerializationVersion = 1;
 
         System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator()
         {
@@ -8644,6 +8694,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                 moduleFileIndex = Microsoft.Diagnostics.Tracing.Etlx.ModuleFileIndex.Invalid;
                 methodOrProcessOrIlMapIndex = -2 - ((int)processIndex);      // Encode process index to make it unambiguous with a method index.
                 InclusiveCount = 0;
+                optimizationTier = Parsers.Clr.OptimizationTier.Unknown;
                 Debug.Assert(GetProcessIndex(null) == processIndex);
             }
 
@@ -8719,7 +8770,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                 TraceLoadedModule loadedModule = process.LoadedModules.FindModuleAndIndexContainingAddress(Address, long.MaxValue - 1, out index);
                 if (loadedModule != null)
                 {
-                    moduleFileIndex = loadedModule.ModuleFile.ModuleFileIndex;
+                    SetModuleFileIndex(loadedModule.ModuleFile);
                     methodOrProcessOrIlMapIndex = -1;           //  set it as the invalid method, destroys memory of process we are in.  
                     return Microsoft.Diagnostics.Tracing.Etlx.MethodIndex.Invalid;
                 }
@@ -8767,6 +8818,32 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                 return moduleFileIndex;
             }
 
+            internal void SetModuleFileIndex(TraceModuleFile moduleFile)
+            {
+                if (moduleFileIndex != Etlx.ModuleFileIndex.Invalid)
+                {
+                    return;
+                }
+
+                moduleFileIndex = moduleFile.ModuleFileIndex;
+
+                if (optimizationTier == Parsers.Clr.OptimizationTier.Unknown &&
+                    moduleFile.IsReadyToRun &&
+                    moduleFile.ImageBase <= Address &&
+                    Address < moduleFile.ImageEnd)
+                {
+                    optimizationTier = Parsers.Clr.OptimizationTier.ReadyToRun;
+                }
+            }
+
+            internal void SetOptimizationTier(OptimizationTier value)
+            {
+                if (optimizationTier == Parsers.Clr.OptimizationTier.Unknown)
+                {
+                    optimizationTier = value;
+                }
+            }
+
             // keep track of how popular each code stack is.  
             internal void UpdateStats()
             {
@@ -8785,6 +8862,8 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
             internal int methodOrProcessOrIlMapIndex;
 
             internal ModuleFileIndex moduleFileIndex;
+
+            internal OptimizationTier optimizationTier;
         }
 
         private ILMapIndex UnloadILMapForMethod(MethodIndex methodIndex, MethodLoadUnloadVerboseTraceData data)
@@ -8999,7 +9078,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
         private GrowableArray<ILToNativeMap> ILToNativeMaps;                    // only Jitted code has these, indexed by ILMapIndex 
         private Dictionary<long, ILMapIndex> methodIDToILToNativeMap;
 
-        private TraceCodeAddress[] codeAddressObjects;  // If we were asked for TraceCodeAddresses (instead of indexes) we cache them
+        private TraceCodeAddress[][] codeAddressObjects;  // If we were asked for TraceCodeAddresses (instead of indexes) we cache them, in sparse array
         private string[] names;                         // A cache (one per code address) of the string name of the address
         private int managedMethodRecordCount;           // Remembers how many code addresses are managed methods (currently not serialized)
         internal int totalCodeAddresses;                 // Count of the number of times a code address appears in the log.
@@ -9497,6 +9576,19 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
             this.methodIndex = methodIndex;
         }
 
+        /// <summary>
+        /// Returns a new string prefixed with the optimization tier if it would be useful. Typically used to adorn a method's
+        /// name with the optimization tier of the specific code version of the method.
+        /// </summary>
+        internal static string PrefixOptimizationTier(string str, OptimizationTier optimizationTier)
+        {
+            if (optimizationTier == OptimizationTier.Unknown || string.IsNullOrWhiteSpace(str))
+            {
+                return str;
+            }
+            return $"[{optimizationTier}]{str}";
+        }
+
         private TraceMethods methods;
         private MethodIndex methodIndex;
         #endregion
@@ -9797,6 +9889,11 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
         public int ImageId { get { return timeDateStamp; } }
 
         /// <summary>
+        /// Tells if the module file is ReadyToRun (the has precompiled code for some managed methods)
+        /// </summary>
+        public bool IsReadyToRun { get { return isReadyToRun; } }
+
+        /// <summary>
         /// If the Product Version fields has a GIT Commit Hash component, this returns it,  Otherwise it is empty.   
         /// </summary>
         public string GitCommitHash
@@ -9876,6 +9973,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                     "PdbSignature=" + XmlUtilities.XmlQuote(PdbSignature) + " " +
                     "PdbAge=" + XmlUtilities.XmlQuote(PdbAge) + " " +
                     "FileVersion=" + XmlUtilities.XmlQuote(FileVersion) + " " +
+                    "IsReadyToRun=" + XmlUtilities.XmlQuote(IsReadyToRun) + " " +
                    "/>";
         }
         #region Private
@@ -9898,6 +9996,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
         internal Address imageBase;
         internal string name;
         private ModuleFileIndex moduleFileIndex;
+        internal bool isReadyToRun;
         internal TraceModuleFile next;          // Chain of modules that have the same path (But different image bases)
 
         internal string pdbName;
@@ -10379,6 +10478,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
         /// If errors occur during conversion, just assume the traced ended at that point and continue. 
         /// </summary>
         public bool ContinueOnError;
+
         #region private
         private TextWriter m_ConversionLog;
         #endregion
